eliminate dates from filenames
Today I realized that having the dates in the filenames makes it
difficult to update articles.  If I refresh an article, I need to rename
it, which breaks the old links.  So, I decided to go ahead and break all
the links now, in order to use names that don't have to be changed in
the future.